Move game specific files in a Game folder
