+++ src/Game/Scene.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <raylib.h>
+#include "raylib.h"
 
 namespace Redge
 {
