+++ src/Tiled/Object.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Scene.h"
+#include "Game/Scene.h"
 
 #include <memory>
